CDS: Retry remote front-end transactions on AskTimeoutException
With the front-end PrimaryShardInfo cache, if the cached primary/leader
shard is remote and unavailable, the RemoteTransactionContextSupport will
fail with an AskTimeoutException when it tries to send the
CreateTransaction message. Since it can take at least 1 election timeout
period to re-elect a new leader, I changed
RemoteTransactionContextSupport to also retry on AskTimeoutException (it
already retries on NoShardLeaderException). However instead of
re-sending the CreateTransaction message, as it did before, it now
re-sends the FindPrimary message to get a new primary shard actor.

I also modified how RemoteTransactionContextSupport retries. It will now
retry for a total period of 2 times the shard election timeout which should
be  ample time for a re-election to occur. If no leader is found then the
txn will fail.

I also added a ShardLeaderNotRespondingException which the
RemoteTransactionContextSupport will throw if it ends up with an
AskTimeoutException after the tx creation timeout period. This shouldn't
occur normally as, with the retries, it should get a NoShardLeaderException
even if the initial error was AskTimeoutException. But it's possible to
end up with an AskTimeoutException, eg if the system is overloaded and
the election timeout is delayed.

For the unit tests, I needed to be able to set the operationTimeout in
millis so that required changes in the DatastoreContext and other
classes.

+++ opendaylight/md-sal/sal-distributed-datastore/src/main/java/org/opendaylight/controller/cluster/datastore/OperationLimiter.java
@@ -28,7 +28,7 @@ public class OperationLimiter extends OnComplete<Object> {
     private final Semaphore semaphore;
     private final int maxPermits;
 
-    OperationLimiter(final TransactionIdentifier identifier, final int maxPermits, final int acquireTimeoutSeconds) {
+    OperationLimiter(final TransactionIdentifier identifier, final int maxPermits, final long acquireTimeoutSeconds) {
         this.identifier = Preconditions.checkNotNull(identifier);
 
         Preconditions.checkArgument(acquireTimeoutSeconds >= 0);
